renaming ValidateParameters.tests.ts -> ValidateParametersTests.ts
